WirdBackend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
asgiref==3.10.0
async-timeout==5.0.1
attrs==25.4.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
defusedxml==0.7.1
dj-rest-auth==7.0.1
Django==5.2.8
django-allauth==65.10.0
django-cachalot==2.8.0
django-cors-headers==4.9.0
django-filter==25.2
django-permissions-policy==4.28.0
django-polymorphic==4.1.0
django-ratelimit==4.1.0
django-resized==1.0.3
django-rest-polymorphic==0.1.10
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-auth-kit==0.3.10
drf-nested-routers==0.95.0
drf-spectacular==0.29.0
drf-yasg==1.21.11
frozenlist==1.8.0
funcy==2.0
hiredis==3.3.0
idna==3.11
inflection==0.5.1
Jinja2==3.1.6
jsonpickle==4.1.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
multidict==6.7.0
oauthlib==3.3.1
packaging==25.0
pillow==12.0.0
pip-review==1.3.0
ply==3.11
propcache==0.4.1
psycopg2-binary==2.9.11
pycparser==2.23
PyJWT==2.10.1
pyotp==2.9.0
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.3
redis==7.0.1
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.28.0
six==1.17.0
sqlparse==0.5.3
stone==3.3.9
structlog==25.5.0
tinydb==4.8.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.11.0
yarl==1.22.0
django-q2>=1.6.0
firebase-admin>=6.0.0
pytz>=2023.3